refactor: move Authority to enum folder
