+++ src/main/java/com/cewar/config/SecurityConfig.java
@@ -18,8 +18,8 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 
+import com.cewar.enums.Authority;
 import com.cewar.services.MyUserDetailsService;
-import com.cewar.web.userdetails.Authority;
 
 @Configuration
 @EnableWebSecurity
