+++ src/main/java/com/cewar/controllers/UserApiController.java
@@ -4,9 +4,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.cewar.dtos.CardDto;
+import com.cewar.enums.Authority;
 import com.cewar.repositories.CardRepository;
 import com.cewar.repositories.UserRepository;
-import com.cewar.web.userdetails.Authority;
 import com.cewar.web.userdetails.User;
 import com.cewar.web.userdetails.UserCard;
